Reorganize a bunch of components
