+++ packages/components/src/ClippedGraph/ClippedGraph.stories.tsx
