+++ packages/components/src/ClippedGraph/ClippedGraph.tsx
