+++ packages/components/src/CollapsiblePanel/CollapsiblePanel.stories.tsx
