+++ packages/components/src/CollapsiblePanel/CollapsiblePanel.types.ts
