--- packages/components/src/CollapsiblePanel/index.tsx
@@ -1,6 +1,0 @@
-/*!
- * Copyright (c) Microsoft. All rights reserved.
- * Licensed under the MIT license. See LICENSE file in the project.
- */
-export * from './CollapsiblePanel.js'
-export * from './CollapsiblePanelContainer.js'
