+++ packages/components/src/SearchBox/SearchBox.tsx
