+++ packages/components/src/Settings/BooleanControl.tsx
@@ -5,8 +5,8 @@
 import { Checkbox, Toggle } from '@fluentui/react'
 import { useCallback } from 'react'
 
-import type { ControlProps } from './interfaces.js'
-import { ControlType } from './interfaces.js'
+import type { ControlProps } from './Settings.types.js'
+import { ControlType } from './Settings.types.js'
 
 /**
  * BooleanControl creates either Checkbox or Toggle Fluent component
