+++ packages/components/src/index.ts
@@ -2,18 +2,26 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-export * from './ButtonChoiceGroup/index.js'
-export * from './ClippedGraph/index.js'
-export * from './CollapsiblePanel/index.js'
+export * from './ButtonChoiceGroup/ButtonChoiceGroup.js'
+export * from './ClippedGraph/ClippedGraph.js'
+export * from './CollapsiblePanel/CollapsiblePanel.js'
+export * from './CollapsiblePanel/CollapsiblePanel.types.js'
+export * from './CollapsiblePanel/CollapsiblePanelContainer.js'
 export * from './ColumnarMenu/index.js'
-export * from './ControlledHistogramFilter/index.js'
-export * from './DialogConfirm/index.js'
+export * from './ControlledHistogramFilter/ControlledHistogramFilter.js'
+export * from './DialogConfirm/DialogConfirm.js'
 export * from './EnumDropdown/EnumDropdown.js'
+export * from './EnumDropdown/EnumDropdown.types.js'
 export * from './Expando/Expando.js'
-export * from './FilterTextbox/index.js'
+export * from './Expando/Expando.types.js'
+export * from './FilterTextbox/FilterTextbox.js'
+export * from './FilterTextbox/FilterTextbox.types.js'
 export * from './MultiDropdown/MultiDropdown.js'
-export * from './NumberSpinButton/index.js'
+export * from './MultiDropdown/MultiDropdown.types.js'
+export * from './NumberSpinButton/NumberSpinButton.js'
 export * from './PolicyAndCookieBanner/index.js'
-export * from './SearchBox/index.js'
-export * from './Settings/index.js'
-export * from './ToggleLink/index.js'
+export * from './SearchBox/SearchBox.js'
+export * from './Settings/Settings.js'
+export type { SettingsGroup, SettingsProps } from './Settings/Settings.types.js'
+export * from './ToggleLink/ToggleLink.js'
+export * from './util/enums.js'
